Fix v1 bug, move v2 plugin to apm-sdk-plugin, modify witnessClasses auto select plugin
