+++ apm-sniffer/apm-sdk-plugin/kafka-v1-plugin/src/main/java/org/apache/skywalking/apm/plugin/kafka/v1/ConsumerConstructorInterceptor.java
@@ -29,6 +29,7 @@ public class ConsumerConstructorInterceptor implements InstanceConstructorInterc
         // set the bootstrap server address
         ConsumerEnhanceRequiredInfo requiredInfo = new ConsumerEnhanceRequiredInfo();
         requiredInfo.setBrokerServers(config.getList("bootstrap.servers"));
+        requiredInfo.setGroupId(config.getString("group.id"));
         objInst.setSkyWalkingDynamicField(requiredInfo);
     }
 }
