+++ apm-sniffer/apm-sdk-plugin/kafka-v1-plugin/src/main/java/org/apache/skywalking/apm/plugin/kafka/v1/define/AbstractKafkaInstrumentation.java
@@ -23,6 +23,6 @@ import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.ClassInst
 public abstract class AbstractKafkaInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
 
     @Override protected String[] witnessClasses() {
-        return new String[]{"org.apache.kafka.clients.ApiVersions"};
+        return new String[]{"org.apache.kafka.clients.consumer.internals.StaleMetadataException"};
     }
 }
