+++ apm-sniffer/apm-sdk-plugin/kafka-v1-plugin/src/test/java/org/apache/skywalking/apm/plugin/kafka/v11/CallbackInterceptorTest.java
@@ -143,6 +143,6 @@ public class CallbackInterceptorTest {
     }
 
     private void assertCallbackSpan(AbstractTracingSpan span) {
-        assertThat(span.getOperationName(), is("Producer/Callback"));
+        assertThat(span.getOperationName(), is("Kafka/null/0/Producer/Callback"));
     }
 }
