+++ apm-sniffer/apm-sdk-plugin/kafka-v2-plugin/src/main/java/org/apache/skywalking/apm/plugin/kafka/v2/ConsumerEnhanceRequiredInfo.java
