+++ apm-sniffer/apm-sdk-plugin/kafka-v2-plugin/src/main/java/org/apache/skywalking/apm/plugin/kafka/v2/define/AbstractKafkaInstrumentation.java
@@ -25,7 +25,8 @@ import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.ClassInst
  */
 public abstract class AbstractKafkaInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
 
-    @Override protected String[] witnessClasses() {
-        return new String[]{"org.apache.kafka.clients.ApiVersions"};
+    @Override
+    protected String[] witnessClasses() {
+        return new String[]{"org.apache.kafka.clients.admin.internals.AdminMetadataManager"};
     }
 }
