+++ apm-sniffer/apm-sdk-plugin/kafka-v2-plugin/src/main/resources/skywalking-plugin.def
