+++ apm-sniffer/apm-sdk-plugin/kafka-v2-plugin/src/test/java/org/apache/skywalking/apm/plugin/kafka/v2/ProducerConstructorInterceptorTest.java
@@ -16,14 +16,12 @@
  *
  */
 
-package org.apache.skywalking.apm.plugin.kafka.v20;
+package org.apache.skywalking.apm.plugin.kafka.v2;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
+import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
-import org.apache.skywalking.apm.plugin.kafka.v2.ConsumerEnhanceRequiredInfo;
-import org.apache.skywalking.apm.plugin.kafka.v2.SubscribeMethodInterceptor;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -32,38 +30,40 @@ import org.mockito.runners.MockitoJUnitRunner;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
+import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class SubscribeMethodInterceptorTest {
+public class ProducerConstructorInterceptorTest {
+    @Mock
+    private ProducerConfig producerConfig;
 
     @Mock
-    private SubscribeMethodInterceptor constructorInterceptor;
-
-    private List<String> mockTopics = new ArrayList<String>();
+    private ProducerConstructorInterceptor constructorInterceptor;
 
     private EnhancedInstance enhancedInstance = new EnhancedInstance() {
-        ConsumerEnhanceRequiredInfo consumerEnhanceRequiredInfo = new ConsumerEnhanceRequiredInfo();
+        private String brokerServers;
 
         @Override public Object getSkyWalkingDynamicField() {
-            return consumerEnhanceRequiredInfo;
+            return brokerServers;
         }
 
         @Override public void setSkyWalkingDynamicField(Object value) {
-            this.consumerEnhanceRequiredInfo = (ConsumerEnhanceRequiredInfo)value;
+            brokerServers = (String)value;
         }
     };
 
     @Before
     public void setUp() {
-        mockTopics.add("test");
-        mockTopics.add("test-1");
-        constructorInterceptor = new SubscribeMethodInterceptor();
+        List<String> mockBootstrapServers = new ArrayList<String>();
+        mockBootstrapServers.add("localhost:9092");
+        mockBootstrapServers.add("localhost:19092");
+        when(producerConfig.getList("bootstrap.servers")).thenReturn(mockBootstrapServers);
+        constructorInterceptor = new ProducerConstructorInterceptor();
     }
 
     @Test
-    public void testOnConsumer() throws Throwable {
-        constructorInterceptor.beforeMethod(enhancedInstance, null, new Object[] {mockTopics}, new Class[] {Collection.class}, null);
-        ConsumerEnhanceRequiredInfo requiredInfo = (ConsumerEnhanceRequiredInfo)enhancedInstance.getSkyWalkingDynamicField();
-        assertThat(requiredInfo.getTopics(), is("test;test-1"));
+    public void testOnConsumer() {
+        constructorInterceptor.onConstruct(enhancedInstance, new Object[] {producerConfig});
+        assertThat(enhancedInstance.getSkyWalkingDynamicField().toString(), is("localhost:9092;localhost:19092"));
     }
 }
