+++ apm-sniffer/apm-sdk-plugin/kafka-v2-plugin/src/test/java/org/apache/skywalking/apm/plugin/kafka/v2/SubscribeMethodInterceptorTest.java
@@ -16,10 +16,12 @@
  *
  */
 
-package org.apache.skywalking.apm.plugin.kafka.v20;
+package org.apache.skywalking.apm.plugin.kafka.v2;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
-import org.apache.skywalking.apm.plugin.kafka.v2.ProducerRecordConstructorInterceptor;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -30,31 +32,36 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 
 @RunWith(MockitoJUnitRunner.class)
-public class ProducerRecordConstructorInterceptorTest {
+public class SubscribeMethodInterceptorTest {
 
     @Mock
-    private ProducerRecordConstructorInterceptor constructorInterceptor;
+    private SubscribeMethodInterceptor constructorInterceptor;
+
+    private List<String> mockTopics = new ArrayList<String>();
 
     private EnhancedInstance enhancedInstance = new EnhancedInstance() {
-        private String brokerServers;
+        ConsumerEnhanceRequiredInfo consumerEnhanceRequiredInfo = new ConsumerEnhanceRequiredInfo();
 
         @Override public Object getSkyWalkingDynamicField() {
-            return brokerServers;
+            return consumerEnhanceRequiredInfo;
         }
 
         @Override public void setSkyWalkingDynamicField(Object value) {
-            brokerServers = (String)value;
+            this.consumerEnhanceRequiredInfo = (ConsumerEnhanceRequiredInfo)value;
         }
     };
 
     @Before
     public void setUp() {
-        constructorInterceptor = new ProducerRecordConstructorInterceptor();
+        mockTopics.add("test");
+        mockTopics.add("test-1");
+        constructorInterceptor = new SubscribeMethodInterceptor();
     }
 
     @Test
-    public void testOnConsumer() {
-        constructorInterceptor.onConstruct(enhancedInstance, new Object[] {"test"});
-        assertThat(enhancedInstance.getSkyWalkingDynamicField().toString(), is("test"));
+    public void testOnConsumer() throws Throwable {
+        constructorInterceptor.beforeMethod(enhancedInstance, null, new Object[] {mockTopics}, new Class[] {Collection.class}, null);
+        ConsumerEnhanceRequiredInfo requiredInfo = (ConsumerEnhanceRequiredInfo)enhancedInstance.getSkyWalkingDynamicField();
+        assertThat(requiredInfo.getTopics(), is("test;test-1"));
     }
 }
