+++ apm-sniffer/optional-plugins/pom.xml
@@ -48,7 +48,6 @@
         <module>zookeeper-3.4.x-plugin</module>
         <module>customize-enhance-plugin</module>
         <module>play-2.x-plugin</module>
-        <module>kafka-v2-plugin</module>
     </modules>
 
     <dependencies>
